# docs/openapi.yaml
openapi: 3.0.3
info:
  title: CERBERUS Commerce OS — API Sözleşmesi (T3.5)
  version: "3.0-faz3"
  description: >
    Tüm uçlar `cerberus_session` imzalı JWT çerezi ile korunur (açık uçlar: /api/auth/login, /api/health).
    Kimliksiz istek 401; rol/storeCode ihlali 403; doğrulama hatası 422; sunucu hatası 500 + correlationId döner.
servers:
  - url: /
...
paths:
...
  /api/auth/login:
    post:
      tags: [auth]
      summary: Giriş — bcrypt doğrulama, rate-limit (5/15dk), imzalı JWT çerez
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email, maxLength: 254 }
                password: { type: string, minLength: 1, maxLength: 128 }
      responses:
        "200": { description: "success + kullanıcı özeti; Set-Cookie: cerberus_session (httpOnly, 8s)" }
        "401": { description: "E-posta/parola hatalı veya iptal edilmiş parola" }
        "422": { description: "Zod doğrulama hatası" }
        "429": { description: "Rate limit; Retry-After başlığı" }
...
  /api/orders/{id}:
    patch:
      tags: [orders]
      summary: Sipariş güncelle (beyaz liste alanlar; storeCode kapsam kontrolü)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: "message + order" }
        "401": { description: Oturum gerekli }
        "403": { description: Mağaza kapsamı dışı }
        "422": { description: Doğrulama hatası }
    delete:
      tags: [orders]
      summary: Sipariş sil (yalnız ADMIN/MANAGER)
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: silindi }
        "401": { description: Oturum gerekli }
        "403": { description: Yetki yetersiz }
...
  /api/admin/database-reset:
    post:
      tags: [admin]
      summary: "Yıkıcı sıfırlama araçları — YALNIZCA geliştirme ortamı (üretimde 404), ADMIN + RESET-CERBERUS onayı"
      responses:
        "200": { description: başarılı }
        "404": { description: Üretimde kapalı }